requirements.txt
torch
torchvision
torchaudio
transformers
streamlit
pillow
opencv-python
